+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Some entries include illustrative screenshots in nested sub-lists.
 
+## [Unreleased]
+
+### Fixed
+
+- Code comments gutter button is now correctly positioned on the hovered line
+
 ## [1.5.0] - 2026-04-20
 
 ### Added